adjust sqlite query to yield total play time during period, fix #94

+++ composeApp/src/androidMain/kotlin/it/fast4x/rimusic/service/PlayerMediaBrowserService.kt
@@ -487,7 +487,7 @@ class PlayerMediaBrowserService : MediaBrowserServiceCompat(), ServiceConnection
                                 .findSongsMostPlayedBetween(
                                     from = 0,
                                     limit = preferences.getEnum( MaxTopPlaylistItemsKey, MaxTopPlaylistItems.`10` )
-                                                       .toLong()
+                                                       .toInt()
                                 )
                                 .first()
